+++ src/Concerns/WithGroupedHeaders.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Maatwebsite\Excel\Concerns;
-
-interface WithGroupedHeaders extends WithHeadingRow
-{
-}