+++ src/Concerns/WithGroupedHeadingRow.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Maatwebsite\Excel\Concerns;
+
+interface WithGroupedHeadingRow extends WithHeadingRow
+{
+}