+++ src/Row.php
@@ -35,11 +35,13 @@ class Row implements ArrayAccess
     /**
      * @param  SpreadsheetRow  $row
      * @param  array  $headingRow
+     * @param  array  $headerIsGrouped
      */
-    public function __construct(SpreadsheetRow $row, array $headingRow = [])
+    public function __construct(SpreadsheetRow $row, array $headingRow = [], array $headerIsGrouped = [])
     {
-        $this->row        = $row;
-        $this->headingRow = $headingRow;
+        $this->row             = $row;
+        $this->headingRow      = $headingRow;
+        $this->headerIsGrouped = $headerIsGrouped;
     }
 
     /**
@@ -82,7 +84,11 @@ class Row implements ArrayAccess
             $value = (new Cell($cell))->getValue($nullValue, $calculateFormulas, $formatData);
 
             if (isset($this->headingRow[$i])) {
-                $cells[$this->headingRow[$i]] = $value;
+                if (!$this->headerIsGrouped[$i]) {
+                    $cells[$this->headingRow[$i]] = $value;
+                } else {
+                    $cells[$this->headingRow[$i]][] = $value;
+                }
             } else {
                 $cells[] = $value;
             }